refactor: modularise
